+++ src/Command/MonitoringMetierCommand.php
@@ -38,15 +38,16 @@ class MonitoringMetierCommand extends Command
 
     public function execute(InputInterface $input, OutputInterface $output): int
     {
-        $date_status_cible = date("c", strtotime("now -15 minutes"));
-
+        $date_status_cible = date("c", strtotime("now -1 minutes"));
 
         $nbPostesParMin = $this->actesTransactionsSQL->getNbPostesDepuis($date_status_cible);
-        list($nbEnAttenteParMin,$delaiPosteEnAttente) = $this->actesTransactionsSQL->getStatusTransitionStatistics(
-            $date_status_cible,
-            ActesStatusSQL::STATUS_POSTE,
-            ActesStatusSQL::STATUS_EN_ATTENTE_DE_TRANSMISSION
-        );
+
+        list($nbEnAttenteParMin,$delaiPosteEnAttente) =
+            $this->actesTransactionsSQL->getStatusTransitionStatistics(
+                $date_status_cible,
+                ActesStatusSQL::STATUS_POSTE,
+                ActesStatusSQL::STATUS_EN_ATTENTE_DE_TRANSMISSION
+            );
         list($nbTransmisParMin,$delaiEnAttenteTransmis) = $this->actesTransactionsSQL->getStatusTransitionStatistics(
             $date_status_cible,
             ActesStatusSQL::STATUS_EN_ATTENTE_DE_TRANSMISSION,
@@ -59,35 +60,44 @@ class MonitoringMetierCommand extends Command
         );
 
 
-        $nbHeliosPostesPar15Min = $this->heliosTransactionsSQL->getNbPostesDepuis($date_status_cible);
-        list($nbHeliosEnAttentePar15Min,$delaiHeliosPosteEnAttente) = $this->heliosTransactionsSQL->getStatusTransitionStatistics(
+        list($nbHeliosPostesPar15Min,$volumePoste) = $this->heliosTransactionsSQL->getNbPostesDepuis($date_status_cible);
+
+        list($nbHeliosEnAttentePar15Min,$delaiHeliosPosteEnAttente,$volumeEnAttente) = $this->heliosTransactionsSQL->getStatusTransitionStatistics(
             $date_status_cible,
             HeliosTransactionsSQL::POSTE,
             HeliosTransactionsSQL::ATTENTE
         );
-        list($nbHeliosTransmisPar15Min,$delaiHeliosEnAttenteTransmis) = $this->heliosTransactionsSQL->getStatusTransitionStatistics(
+
+
+        list($nbHeliosTransmisPar15Min,$delaiHeliosEnAttenteTransmis,$volumeTransmis) = $this->heliosTransactionsSQL->getStatusTransitionStatistics(
             $date_status_cible,
             HeliosTransactionsSQL::ATTENTE,
             HeliosTransactionsSQL::TRANSMIS
         );
-        list($nbHeliosAcquittementPar15Min,$delaiHeliosTransmisAcquittement) = $this->heliosTransactionsSQL->getStatusTransitionStatistics(
+
+        list($nbHeliosAcquittementPar15Min,$delaiHeliosTransmisAcquittement,$volumeAcquite) = $this->heliosTransactionsSQL->getStatusTransitionStatistics(
             $date_status_cible,
             HeliosTransactionsSQL::TRANSMIS,
             HeliosTransactionsSQL::INFORMATION_DISPONIBLE
         );
 
         $resultats = [
-            "nbPostesParMin" => $nbPostesParMin / 15,
-            "nbEnAttenteParMin" => $nbEnAttenteParMin / 15,
-            "nbTransmisParMin" => $nbTransmisParMin / 15,
-            "nbAcquittesParMin" => $nbAcquittementParMin / 15,
+            "nbPostesParMin" => $nbPostesParMin,
+            "nbEnAttenteParMin" => $nbEnAttenteParMin,
+            "nbTransmisParMin" => $nbTransmisParMin,
+            "nbAcquittesParMin" => $nbAcquittementParMin,
             "delaiPosteEnAttente" => $delaiPosteEnAttente,
             "delaiEnAttenteTransmis" => $delaiEnAttenteTransmis,
             "delaiTransmisAcquittement" => $delaiTransmisAcquittement,
-            "nbHeliosPostesParMin" => $nbHeliosPostesPar15Min / 15,
-            "nbHeliosEnAttenteParMin" => $nbHeliosEnAttentePar15Min / 15,
-            "nbHeliosTransmisParMin" => $nbHeliosTransmisPar15Min / 15,
-            "nbHeliosAcquittesParMin" => $nbHeliosAcquittementPar15Min / 15,
+            "nbHeliosPostesParMin" => $nbHeliosPostesPar15Min,
+
+            "nbHeliosEnAttenteParMin" => $nbHeliosEnAttentePar15Min,
+            "nbHeliosTransmisParMin" => $nbHeliosTransmisPar15Min,
+            "nbHeliosAcquittesParMin" => $nbHeliosAcquittementPar15Min,
+            "volumePosteParMin" => $volumePoste,
+            "volumeEnAttenteParMin" => $volumeEnAttente,
+            "volumeTransmisParMin" => $volumeTransmis,
+
             "delaiHeliosPosteEnAttente" => $delaiHeliosPosteEnAttente,
             "delaiHeliosEnAttenteTransmis" => $delaiHeliosEnAttenteTransmis,
             "delaiHeliosTransmisAcquittement" => $delaiHeliosTransmisAcquittement,